refactor: KakaoMap 컴포넌트 공통 ui 폴더로 이동
